% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixgaussian.R
\name{mixgaussian}
\alias{mixgaussian}
\title{Mixture of two Gaussian functions.}
\usage{
mixgaussian(x, H1, H2, m1, m2, s1, s2)
}
\arguments{
\item{x}{Environmental variable}

\item{H1}{Height of the first Gaussian}

\item{H2}{Height of the second Gaussian}

\item{m1}{Mode of the first Gaussian}

\item{m2}{Mode of the second Gaussian}

\item{s1}{Spread of the first Gaussian}

\item{s2}{Spread of the second Gaussian}
}
\value{
Value at x, given the other parameters
}
\description{
Mixture of two Gaussian functions.
}
