% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixgaussian.R
\docType{data}
\name{mixgaussian_stan}
\alias{mixgaussian_stan}
\title{Mixture of two Gaussian functions Stan code}
\description{
Stan code for the mixgaussian function
}
\examples{
cat(mixgaussian_stan[["source_code"]])
}
\keyword{datasets}
