% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modskurt.R
\name{modskurt}
\alias{modskurt}
\title{Full modskurt}
\usage{
modskurt(x, H, m, s, r, d, p)
}
\arguments{
\item{x}{Environmental variable}

\item{H}{Height of the mode}

\item{m}{Location of the mode}

\item{s}{Spread}

\item{r}{Skewness}

\item{d}{Flatness}

\item{p}{Pointyness}
}
\value{
Value at x, given the other parameters.
}
\description{
Full modskurt
}
