% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modskurt.R
\docType{data}
\name{modskurt_stan}
\alias{modskurt_stan}
\title{Full modskurt Stan code}
\description{
Stan code for the full modskurt function
}
\examples{
cat(modskurt_stan[["source_code"]])
}
\keyword{datasets}
