% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modskurt.R
\name{modskurtd0}
\alias{modskurtd0}
\title{Modskurt with asymmetry allowing "exaggeration" / large tails}
\usage{
modskurtd0(x, H, m, s, r, p)
}
\arguments{
\item{x}{Environmental variable}

\item{H}{Height of the mode}

\item{m}{Location of the mode}

\item{s}{Spread}

\item{r}{Skewness}

\item{p}{Pointyness}
}
\value{
Value at x, given the other parameters.
}
\description{
Modskurt with asymmetry allowing "exaggeration" / large tails
}
