% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modskurt.R
\docType{data}
\name{modskurtd0_stan}
\alias{modskurtd0_stan}
\title{Modskurt with asymmetry allowing "exaggeration" / large tails Stan code.}
\description{
Stan code for the modskurt with asymmetry allowing "exaggeration" / large tails.
}
\examples{
cat(modskurtd0_stan[["source_code"]])
}
\keyword{datasets}
