% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modskurt.R
\name{modskurtp1}
\alias{modskurtp1}
\title{Asymetric with varied flattening / peakedness}
\usage{
modskurtp1(x, H, m, s, r, d)
}
\arguments{
\item{x}{Environmental variable}

\item{H}{Height of the mode}

\item{m}{Location of the mode}

\item{s}{Spread}

\item{r}{Skewness}

\item{d}{Flatness}
}
\value{
Value at x, given the other parameters.
}
\description{
Asymetric with varied flattening / peakedness
}
