% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modskurt.R
\docType{data}
\name{modskurtp1_stan}
\alias{modskurtp1_stan}
\title{Modskurt with Asymmetry and varied flattening / peakedness Stan code.}
\description{
Stan code for the modskurt with asymmetry and varied flattening / peakedness
}
\examples{
cat(modskurtp1_stan[["source_code"]])
}
\keyword{datasets}
