% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modskurt.R
\name{modskurtp1d0}
\alias{modskurtp1d0}
\title{Asymetric Modskurt}
\usage{
modskurtp1d0(x, H, m, s, r)
}
\arguments{
\item{x}{Environmental variable}

\item{H}{Height of the mode}

\item{m}{Location of the mode}

\item{s}{Spread}

\item{r}{Skewness}
}
\value{
Value at x, given the other parameters.
}
\description{
Asymetric Modskurt
}
