% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modskurt.R
\docType{data}
\name{modskurtp1d0_stan}
\alias{modskurtp1d0_stan}
\title{Asymmetric modskurt Stan code.}
\description{
Stan code for the modskurt with asymmetry
}
\examples{
cat(modskurtp1d0_stan[["source_code"]])
}
\keyword{datasets}
