% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modskurt.R
\name{modskurtp1d0r05}
\alias{modskurtp1d0r05}
\title{Symmetric Modskurt}
\usage{
modskurtp1d0r05(x, H, m, s)
}
\arguments{
\item{x}{Environmental variable}

\item{H}{Height of the mode}

\item{m}{Location of the mode}

\item{s}{Spread}
}
\value{
Value at x, given the other parameters.
}
\description{
Symmetric Modskurt
}
