% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modskurt.R
\docType{data}
\name{modskurtp1d0r05_stan}
\alias{modskurtp1d0r05_stan}
\title{Symmetric Modskurt Stan code.}
\description{
Stan code for the symmetric Modskurt.
}
\examples{
cat(modskurtp1d0r05_stan[["source_code"]])
}
\keyword{datasets}
