% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modskurt.R
\name{modskurtp1r05}
\alias{modskurtp1r05}
\title{Symmetric Modskurt with varying flattening / peakedness}
\usage{
modskurtp1r05(x, H, m, s, d)
}
\arguments{
\item{x}{Environmental variable}

\item{H}{Height of the mode}

\item{m}{Location of the mode}

\item{s}{Spread}

\item{d}{Flatness}
}
\value{
Value at x, given the other parameters.
}
\description{
Symmetric Modskurt with varying flattening / peakedness
}
