% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modskurt.R
\docType{data}
\name{modskurtp1r05_stan}
\alias{modskurtp1r05_stan}
\title{Symmetric Modskurt with varying flattening / peakedness Stan code.}
\description{
Stan code for the symmetric Modskurt with varying flattening / peakedness.
}
\examples{
cat(modskurtp1r05_stan[["source_code"]])
}
\keyword{datasets}
